requirements.txt
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.17.0
pandas>=2.0.0
flask-caching>=2.1.0
python-dotenv>=1.0.0
werkzeug>=2.3.0
gunicorn>=21.2.0
